refactor(fasthttphandler): Comment and  modularize responseWriter and pool methods
